+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notables change for Cocoa Discord Utils will be noted here
 
 **Note 2**: Not all change will be noted, see commits for *full* changelog
 
+## [1.6.0] - 2022-04-21
+
+- Fixed an issues that prevents help command from being generated for slash class
+
+- Validating Commands for Slash Center is now `required`
+
 ## [1.5.0] - 2022-04-21
 
 - Added @FutureSlash for Slash Command that cannot be resolve synchronously